requirements.txt
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.3
responses==0.24.1 